# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Adaptive Math API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /topics:
    get:
      operationId: getTopics
      tags: [topics]
      summary: Get all topics with prerequisite graph
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Topics list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
  /topics/{topicId}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get a single topic
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Topic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
...
  /tutor/hint:
    post:
      operationId: getTutorHint
      tags: [tutor]
      summary: Get a Socratic hint from Stella for a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorHintBody"
      responses:
        "200":
          description: Socratic hint from Stella
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorHintResponse"
  /tutor/chat:
    post:
      operationId: sendTutorMessage
      tags: [tutor]
      summary: Send a message to Stella AI tutor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorChatBody"
      responses:
        "200":
          description: Stella's response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorChatResponse"
...
components:
  schemas:
...
    Topic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        grade:
          type: integer
        category:
          type: string
        prerequisites:
          type: array
          items:
            type: string
          description: List of prerequisite topic IDs
        difficulty:
          type: number
      required:
        - id
        - name
        - grade
        - category
        - prerequisites
        - difficulty
...
    TutorHintBody:
      type: object
      properties:
        questionId:
          type: string
        hintLevel:
          type: integer
          minimum: 1
          maximum: 4
          description: "1=concept, 2=steps, 3=check understanding, 4=full solution"
        studentAnswer:
          type: string
          nullable: true
      required:
        - questionId
        - hintLevel
    TutorHintResponse:
      type: object
      properties:
        hintLevel:
          type: integer
        message:
          type: string
        socraticType:
          type: string
          enum: [CLARIFICATION, ASSUMPTIONS, EVIDENCE, IMPLICATIONS, PERSPECTIVES]
          nullable: true
        followUpQuestion:
          type: string
          nullable: true
      required:
        - hintLevel
        - message
    TutorChatBody:
      type: object
      properties:
        message:
          type: string
        questionId:
          type: string
          nullable: true
        topicId:
          type: string
          nullable: true
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - message
        - conversationHistory
    TutorChatResponse:
      type: object
      properties:
        message:
          type: string
        socraticType:
          type: string
          nullable: true
        suggestedNextStep:
          type: string
          nullable: true
        relatedTopicIds:
          type: array
          items:
            type: string
      required:
        - message
        - relatedTopicIds
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
      required:
        - role
        - content
        - timestamp
...