# lib/api-zod/src/generated/types/topic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adaptive Math API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Topic {
  id: string;
  name: string;
  description?: string;
  grade: number;
  category: string;
  /** List of prerequisite topic IDs */
  prerequisites: string[];
  difficulty: number;
}